# Rewrite Nested Loop Join executor for 3.5× speed and 1% memory usage