@2010YOUY01 2025-07-31
Draft PR and early discussions: https://github.com/apache/datafusion/pull/16889

Thanks @UBarney  @ding-young  and @jonathanc-n for the help 🙏🏼