@ding-young 2025-08-01
### Memory Usage Benchmark
#### current 
<details>

```
Query     Time (ms)     Peak RSS  Peak Commit  Major Page Faults
----------------------------------------------------------------
1            761.84    1042.7 MB       5.9 GB                  0
2            973.85    1486.3 MB       8.3 GB                  0
3           1197.72       2.2 GB      11.9 GB                  0
4           3644.36      12.2 GB      20.8 GB                  0
5           3729.40       3.9 GB      13.4 GB                  0
6          18882.52      22.9 GB      24.5 GB                  0
7           3128.09       3.9 GB      13.4 GB                  0
8          19002.03      22.9 GB      24.6 GB                  0
9           2968.10      10.3 GB      15.1 GB                  0
10          6064.70      19.4 GB      28.3 GB                  0
```

</details>


#### nlj-rewrite
<details>

```
Query     Time (ms)     Peak RSS  Peak Commit  Major Page Faults
----------------------------------------------------------------
1            218.46      78.6 MB    1024.1 MB                  0
2            318.69      76.7 MB    1024.2 MB                  0
3            352.38      77.0 MB    1024.2 MB                  0
4           1384.05      85.1 MB    1024.2 MB                  0
5           2217.99      82.4 MB    1024.2 MB                  0
6           4331.11      77.0 MB    1024.1 MB                  0
7           2295.73      79.1 MB    1024.1 MB                  0
8           4372.36      78.6 MB    1024.1 MB                  0
9            911.73      81.1 MB    1024.2 MB                  0
10          1212.94      71.6 MB    1024.1 MB                  0
```

</details>

(reproducer branch & comments to be added)