@UBarney 2025-08-02
Hi @2010YOUY01 

Thanks for creating this amazing PR and for the detailed explanation on why we don't need to maintain the right_side order. This is a great optimization!

Just to provide some context for future collaboration, this approach (joining only one left row at a time) was actually the initial direction I considered. I was uncertain about the right_side_ordered requirement, so I raised this specific concern and @-mentioned you in [my comment](https://github.com/apache/datafusion/issues/16364#issuecomment-2975520489)  seeking clarification.

Since I didn't get a response, I proceeded with the alternative implementation to keep things moving. A quick clarification on that point back then would have saved the effort on my [merged PR](https://github.com/apache/datafusion/pull/16443) and allowed us to get to this optimal solution directly.

No worries, the main thing is we've landed the best solution for the project. Just thought this feedback on the process could be helpful for our future interactions.

Great work on this!