@2010YOUY01 2025-08-02
@UBarney I'm so sorry about that — I completely let that issue discussion slip through.
For important matters, feel free to ping me multiple times or bring it up again in a PR discussion.
I have to admit I often forget to respond to issue replies, but I usually follow up on PR discussions. (I think probably it's true for others as well)